orders-service/requirements.txt
flask==2.3.3
mysql-connector-python==8.3.0
flask-jwt-extended==4.6.0